# docs/swagger.yaml
basePath: /
definitions:
...
  models.Branch:
    properties:
      aashram_area:
        type: number
      address:
        type: string
      city:
        type: string
      contact_number:
        type: string
      coordinator_name:
        type: string
      country:
        type: string
      created_by:
        type: string
      created_on:
        type: string
      daily_end_time:
        type: string
      daily_start_time:
        type: string
      district:
        type: string
      email:
        type: string
      established_on:
        type: string
      id:
        type: integer
      name:
        type: string
      open_days:
        description: e.g. "Mon-Fri"
        type: string
      pincode:
        type: string
      police_station:
        type: string
      post_office:
        type: string
      state:
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    type: object
...
  models.Event:
    properties:
      event_name:
        type: string
      id:
        type: integer
    type: object
...
  models.Role:
    properties:
      created_on:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_on:
        type: string
    type: object
...
  models.User:
    properties:
      contact_number:
        type: string
      created_by:
        type: string
      created_on:
        type: string
      email:
        type: string
      expired_on:
        type: string
      first_login_on:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      last_login_on:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: integer
      token:
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    type: object
  models.Volunteer:
    properties:
      branch:
        $ref: '#/definitions/models.Branch'
      branch_id:
        type: integer
      created_by:
        type: string
      created_on:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: integer
      id:
        type: integer
      mention_seva:
        type: string
      number_of_days:
        type: integer
      search_volunteer:
        type: string
      seva_involved:
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
      volunteer_name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: vaish@gmail.com
    name: API Support
  description: API documentation for the DJJS Event Reporting Backend.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DJJS Event Reporting API
  version: "1.0"
paths:
...
  /api/events/{event_id}/volunteers:
    get:
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Volunteer'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get volunteer by event ID
      tags:
      - Volunteers
...
  /api/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create user with auto-generated password (returned in response)
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new user
      tags:
      - Users
  /api/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a user (soft delete)
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: user
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a user
      tags:
      - Users
  /api/users/search:
    get:
      description: Retrieve users based on provided filters (email, contact number,
        or both).
      parameters:
      - description: User Email
        in: query
        name: email
        type: string
      - description: User Contact Number
        in: query
        name: contact_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Search users by email or contact number
      tags:
      - Users
  /api/volunteers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Volunteer'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all volunteers
      tags:
      - Volunteers
    post:
      consumes:
      - application/json
      description: Store volunteer details
      parameters:
      - description: Volunteer payload
        in: body
        name: volunteer
        required: true
        schema:
          $ref: '#/definitions/models.Volunteer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a volunteer
      tags:
      - Volunteers
  /api/volunteers/{id}:
    delete:
      parameters:
      - description: Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a volunteer
      tags:
      - Volunteers
    put:
      consumes:
      - application/json
      parameters:
      - description: Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a volunteer
      tags:
      - Volunteers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"